requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
pydantic==2.7.4
pydantic_core==2.18.4
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
SQLAlchemy==2.0.32
typing_extensions==4.12.2
yarl==1.15.1
openpyxl==3.1.5
pandas==2.2.3
